+++ src/aui-char-counter/HISTORY.md
@@ -1,2 +1,7 @@
-aui-char-counter
-========
+# AUI Char Counter
+
+> Documentation and test modifications are not included in this changelog. For more details, see [full commit history](https://github.com/liferay/alloy-ui/commits/master/src/aui-char-counter).
+
+## @VERSION@
+
+* [AUI-1715](https://issues.liferay.com/browse/AUI-1715) Character count for textarea element input in Char Counter is incorrectly calculated in Chrome